Skeleton of DFU app is building
